zopsm/requirements.txt
alembic==0.9.5
astroid==2.2.5
basho-erlastic==2.1.1
certifi==2018.10.15
chardet==3.0.4
credis==1.0.8
Cython==0.29
falcon==1.3.0
future==0.17.1
gevent==1.2.2
graceful==0.5.0
greenlet==0.4.15
gunicorn==19.7.1
hiredis==0.3.0
hvac==0.7.0
idna==2.7
isort==4.3.16
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
pika==0.10.0
ply==3.4
psycopg2==2.7.3.1
pycurl==7.43.0.2
pyhcl==0.2.3
PyJWT==1.5.3
pylint==2.3.1
pyrabbit2==1.0.6
pyresttest==1.7.2.dev0
python-consul==1.1.0
python-dateutil==2.7.5
python-editor==1.0.3
python-logging-rabbitmq==1.0.9
python-mimeparse==1.6.0
python3-protobuf==2.5.0
PyYAML==3.13
redis==2.10.6
requests==2.20.1
riak==2.7.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.14
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
